web/requirements.txt
# Flask и основные зависимости
Flask==2.3.3
Werkzeug==2.3.7

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9  # PostgreSQL драйвер

# Authentication & Security
Flask-JWT-Extended==4.6.0
Authlib==1.2.1  # OAuth библиотека
bcrypt==4.1.1  # Хеширование паролей
python-dotenv==1.0.0  # Загрузка .env файлов

# Validation
marshmallow==3.20.1
Flask-Marshmallow==0.15.0

# HTTP requests
requests==2.31.0

# Payment
yookassa==3.0.0  # ЮKassa SDK

# Email (опционально)
Flask-Mail==0.9.1

# CORS
Flask-CORS==4.0.0

# Utilities
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
